Move get_iac_plugin inside IacProjectResolver
